distance and plot functions
