--- GeoTable.py
@@ -1,79 +1,0 @@
-from datascience import Table
-import pandas as pd
-import geopandas as gpd
-from shapely.geometry import Point
-import geohash2  # type: ignore # pip install geohash2
-
-
-class GeoTable(Table):
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        if 'geometry' not in self.labels:
-            self.append_column('geometry', [None] * self.num_rows)
-        self._geometry = 'geometry'
-
-    
-    def distance_to(self, other, ref_index=0, new_column='distance_to_ref'):
-        """
-        Computes distance in meters from each point in this GeoTable to a reference point
-        in another GeoTable.
-
-        Args:
-            other (GeoTable): Another GeoTable containing reference point.
-            ref_index (int): Index of the reference point in the other GeoTable.
-            new_column (str): Column name to store distances.
-        """
-        if not isinstance(other, GeoTable):
-            raise TypeError("Other must be a GeoTable.")
-
-        gdf_self = self.to_geodataframe().set_crs("EPSG:4326")
-        gdf_other = other.to_geodataframe().set_crs("EPSG:4326")
-
-        if ref_index < 0 or ref_index >= len(gdf_other):
-            raise IndexError(f"Reference index {ref_index} out of bounds.")
-
-        # Reproject to metric CRS for distance calculation
-        gdf_self_proj = gdf_self.to_crs("EPSG:3857")
-        ref_point = gdf_other.to_crs("EPSG:3857").geometry.iloc[ref_index]
-
-        distances = gdf_self_proj.geometry.distance(ref_point)
-        self[new_column] = distances.round(1).tolist()
-
-    def plot(self, zoom=12, **kwargs):
-        """
-        Plot the GeoTable on a basemap using contextily.
-
-        Args:
-            zoom (int): Zoom level for the basemap.
-            **kwargs: Additional arguments passed to GeoDataFrame.plot()
-        """
-        if self._geometry not in self.labels:
-            raise ValueError(f"Geometry column '{self._geometry}' not found.")
-
-        try:
-            import contextily as ctx
-            import matplotlib.pyplot as plt
-        except ImportError:
-            raise ImportError("Install contextily and matplotlib to enable plotting.")
-
-        # Convert to GeoDataFrame with proper CRS
-        gdf = self.to_geodataframe().set_crs("EPSG:4326")
-        gdf_proj = gdf.to_crs("EPSG:3857")  # Web Mercator projection
-
-        # Create the figure
-        fig, ax = plt.subplots(figsize=(10, 8))
-
-        # Plot the points on top of basemap
-        gdf_proj.plot(ax=ax, markersize=20, color='blue', alpha=0.7, edgecolor='white', linewidth=0.5)
-
-        # Add basemap below the points
-        ctx.add_basemap(ax, source=ctx.providers.CartoDB.Positron, zoom=zoom)
-
-        # Clean up plot aesthetics
-        ax.set_axis_off()
-        plt.tight_layout()
-
-        # Display the figure (no return to suppress extra output)
-        plt.show()
-
